Services avancés 1

+++ src/main/java/tn/esprit/tp_foyer/entity/Chambre.java
@@ -19,7 +19,7 @@ public class Chambre implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long idChambre;
 
-    private int numeroChambre;
+    private long numeroChambre;
 
     @Enumerated(EnumType.STRING)
     private TypeChambre TypeC;
